+++ reoptjl/test/test_http_endpoints.py
@@ -109,7 +109,7 @@ class TestHTTPEndpoints(ResourceTestCaseMixin, TestCase):
         # Check the endpoint logic with the expected selection
         self.assertEqual(http_response["prime_mover"], "steam_turbine")
         self.assertEqual(http_response["size_class"], 1)
-        self.assertGreater(http_response["chp_size_based_on_avg_heating_load_kw"], 574.419)
+        self.assertGreater(http_response["chp_elec_size_heuristic_kw"], 574.419)
 
     def test_absorption_chiller_defaults(self):
 
@@ -181,24 +181,95 @@ class TestHTTPEndpoints(ResourceTestCaseMixin, TestCase):
         v2_response = json.loads(resp.content)   
         assert("Error" in v2_response.keys())
 
-    def test_emissions_profile_endpoint(self):
-        # Call to the django view endpoint dev/emissions_profile which calls the http.jl endpoint
+    def test_avert_emissions_profile_endpoint(self):
+        # Call to the django view endpoint dev/avert_emissions_profile which calls the http.jl endpoint
+        #case 1: location in CONUS (Seattle, WA)
         inputs = {
             "latitude": 47.606211,
-            "longitude": -122.336052
+            "longitude": -122.336052,
+            "load_year": 2021
         }
-        resp = self.api_client.get(f'/v3/emissions_profile', data=inputs)
+        resp = self.api_client.get(f'/v3/avert_emissions_profile', data=inputs)
         self.assertHttpOK(resp)
         view_response = json.loads(resp.content)
-        self.assertEquals(view_response["meters_to_region"], 0.0)
-        self.assertEquals(view_response["region"], "Northwest")
+        self.assertEquals(view_response["avert_meters_to_region"], 0.0)
+        self.assertEquals(view_response["avert_region"], "Northwest")
         self.assertEquals(len(view_response["emissions_factor_series_lb_NOx_per_kwh"]), 8760)
+        #case 2: location off shore of NJ (works for AVERT, not Cambium)
         inputs = {
-            "latitude": 47.606211,
-            "longitude": 122.336052
+            "latitude": 39.034417, 
+            "longitude": -74.759292,
+            "load_year": 2021
         }
-        resp = self.api_client.get(f'/v3/emissions_profile', data=inputs)
+        resp = self.api_client.get(f'/v3/avert_emissions_profile', data=inputs)
+        self.assertHttpOK(resp)
+        view_response = json.loads(resp.content)
+        self.assertAlmostEqual(view_response["avert_meters_to_region"], 760.62, delta=1.0)
+        self.assertEquals(view_response["avert_region"], "Mid-Atlantic")
+        self.assertEquals(len(view_response["emissions_factor_series_lb_NOx_per_kwh"]), 8760)
+        #case 3: Honolulu, HI (works for AVERT but not Cambium)
+        inputs = {
+            "latitude": 21.3099, 
+            "longitude": -157.8581,
+            "load_year": 2021
+        }
+        resp = self.api_client.get(f'/v3/avert_emissions_profile', data=inputs)
+        self.assertHttpOK(resp)
+        view_response = json.loads(resp.content)
+        self.assertEquals(view_response["avert_meters_to_region"], 0.0)
+        self.assertEquals(view_response["avert_region"], "Hawaii (Oahu)")
+        self.assertEquals(len(view_response["emissions_factor_series_lb_NOx_per_kwh"]), 8760)
+        #case 4: location well outside of US (does not work)
+        inputs = {
+            "latitude": 0.0,
+            "longitude": 0.0,
+            "load_year": 2022
+        }
+        resp = self.api_client.get(f'/v3/avert_emissions_profile', data=inputs)
         self.assertHttpBadRequest(resp)
+        view_response = json.loads(resp.content)
+        self.assertTrue("error" in view_response)
+        
+
+    def test_cambium_emissions_profile_endpoint(self):
+        # Call to the django view endpoint v3/cambium_emissions_profile which calls the http.jl endpoint
+        #case 1: location in CONUS (Seattle, WA)
+        inputs = {
+            "load_year": 2021,
+            "scenario": "Mid-case",
+            "location_type": "States", 
+            "latitude": 47.606211, # Seattle 
+            "longitude": -122.336052, # Seattle 
+            "start_year": 2024,
+            "lifetime": 25,
+            "metric_col": "lrmer_co2e",
+            "grid_level": "enduse"
+        }
+        resp = self.api_client.get(f'/v3/cambium_emissions_profile', data=inputs) 
+        self.assertHttpOK(resp)
+        view_response = json.loads(resp.content)
+        self.assertEquals(view_response["metric_col"], "lrmer_co2e")
+        self.assertEquals(view_response["location"], "Washington") 
+        self.assertEquals(len(view_response["emissions_factor_series_lb_CO2_per_kwh"]), 8760)
+        #case 2: location off shore of NJ (works for AVERT, not Cambium)
+        inputs["latitude"] = 39.034417
+        inputs["longitude"] = -74.759292
+        resp = self.api_client.get(f'/v3/cambium_emissions_profile', data=inputs) 
+        self.assertHttpBadRequest(resp)
+        view_response = json.loads(resp.content)
+        self.assertTrue("error" in view_response)
+        #case 3: Honolulu, HI (works for AVERT but not Cambium)
+        inputs["latitude"] = 21.3099
+        inputs["longitude"] = -157.8581
+        resp = self.api_client.get(f'/v3/cambium_emissions_profile', data=inputs)
+        self.assertHttpBadRequest(resp) 
+        view_response = json.loads(resp.content)
+        self.assertTrue("error" in view_response)
+        #case 4: location well outside of US (does not work)
+        inputs["latitude"] = 0.0
+        inputs["longitude"] = 0.0
+        resp = self.api_client.get(f'/v3/cambium_emissions_profile', data=inputs)
+        self.assertHttpBadRequest(resp) 
         view_response = json.loads(resp.content)
         self.assertTrue("error" in view_response)